requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
pydantic==2.12.4
scikit-learn==1.7.2
pandas==2.3.3
numpy==2.3.5
joblib==1.5.2
python-dotenv==1.2.1
torch==2.9.1
tqdm==4.67.1
jupyter==1.1.1

# Additional transitive pins (optional, uncomment if strict reproducibility needed)
# starlette==0.50.0
# requests==2.32.5
# click==8.3.1
pytest